% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statement.R
\name{back_quote}
\alias{back_quote}
\title{Back Quote Object}
\usage{
back_quote(x, as_double = NULL)
}
\arguments{
\item{x}{An R object.}

\item{as_double}{Optional. \code{TRUE} or \code{FALSE} which indicates if to
differentiate between type double and integer. The default value is
\code{TRUE}, which means integers are handled as doubles.}
}
\value{
A character vector.
}
\description{
Convert an R object to character and add back quotations.
}
\examples{
back_quote(1:3)
back_quote(1:3, FALSE)
back_quote(NULL)
back_quote(list(c, 1:3, "a"))
}
