% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-string.R
\name{check_string}
\alias{check_string}
\alias{is_string}
\title{Check If Argument Is Single Character}
\usage{
check_string(
  x,
  name = NULL,
  general = NULL,
  specific = NULL,
  supplement = NULL,
  ...
)

is_string(x)
}
\arguments{
\item{x}{The argument to check, which can be any object.}

\item{name}{A single character which gives the argument's name.
The name is used in the error message. By default, the name of the
argument passed to argument \code{x} is captured automatically.}

\item{general}{Optional. A single character which is used to give a
general statement of the error incurred. By default, this is generated
automatically.}

\item{specific}{Optional. A single character which gives a detailed
description of the error. By default, this is generated automatically.}

\item{supplement}{Optional. A (named) character vector which gives some
additional information about the error. The names are used to create
bullets, see \code{\link[=throw]{throw()}}. By default, this is left empty.}

\item{...}{Optional. Additional arguments which can be retrieved with
\code{\link[=tryCatch]{tryCatch()}}.}
}
\value{
returns an invisible \code{NULL} if the argument is valid, or
generates an error message.
}
\description{
Check if an argument is a single character. and if not,
generate an error message.

Can be used to check argument names, for example.
}
\examples{
x <- "a"
check_string(x)

\dontrun{
# `x` must have type character
x <- c
check_string(x)

# `x` must have length 1
x <- c("a", "b")
check_string(x)

# `NA_character_` is not acceptable
x <- NA_character_
check_string(x)
}
}
\seealso{
"Examples" section in \code{\link[=check_type]{check_type()}} for how to customize
error message and how to add and retrieve additional arguments.

\code{vignette("erify")} for a gentle introduction to this package.
}
