% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/type.R
\name{check_type}
\alias{check_type}
\title{Check Argument's Type}
\usage{
check_type(
  x,
  valid,
  name = NULL,
  general = NULL,
  specifics = NULL,
  supplement = NULL,
  ...
)
}
\arguments{
\item{x}{The argument to check, which can be any object.}

\item{valid}{A character vector which contains the valid types.}

\item{name}{A single character which gives the argument's name.
The name is used in the error message. By default, the name of the
argument passed to argument \code{x} is captured automatically.}

\item{general}{Optional. A single character which is used to give a
general statement of the error incurred. This is generated automatically
by default.}

\item{specifics}{Optional. A (named) character vector which gives a
list of detailed descriptions of the error. This is generated automatically
by default.}

\item{supplement}{Optional. A single character which gives some
additional information about the error. This is left empty by default.}

\item{...}{Optional. Additional arguments passed to \code{\link[rlang:abort]{rlang::abort()}},
which is called internally.}
}
\value{
An invisible \code{NULL} if the argument is valid, or an error message
is generated.
}
\description{
Check if an argument has valid type,
and if not, generate an error message.
}
\examples{
# argument to check
arg <- 10

# returns silently if the argument has valid type
check_type(arg, "double")

\dontrun{
check_type(arg, "character")

# specify argument's name
check_type(arg, "character", name = "x")

# customize error message
check_type(
  arg,
  "character",
  general = "General statement of the error.",
  specifics = c(x = "An error.", i = "A hint.", `*` = "Other detail."),
  supplement = "More words to say."
)
}

# add and retrieve additional arguments
tryCatch(
  {check_type(arg, "character", your_arg = "your data")},
  error = function(e) e$your_arg
)
}
\seealso{
\code{\link[=Statement]{Statement()}} for more details about arguments \code{general},
\code{specifics} and \code{supplement}.

\code{\link[rlang:abort]{rlang::abort()}} for adding additional arguments.

\code{vignette("erify")} for a gentle introduction to this package.

\href{https://style.tidyverse.org/error-messages.html}{The tidyverse style guide}
for more details about the used error message style.
}
