% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/type.R
\name{check_types}
\alias{check_types}
\title{Check Each Item's Type}
\usage{
check_types(
  x,
  valid,
  name = NULL,
  general = NULL,
  specific = NULL,
  supplement = NULL,
  n = NULL,
  ...
)
}
\arguments{
\item{x}{The argument to check, which must be a list.}

\item{valid}{A character vector which contains the valid types.}

\item{name}{A single character which gives the argument's name.
The name is used in the error message. By default, the name of the
argument passed to argument \code{x} is captured automatically.}

\item{general}{Optional. A single character which is used to give a
general statement of the error incurred. This is generated automatically
by default.}

\item{specific}{Optional. A single character which gives the template for
reporting each invalid item. R expressions can be inserted with
\code{\link[glue:glue]{glue::glue()}} syntax. See "Details" section for more details. By default,
this argument is generated automatically.}

\item{supplement}{Optional. A single character which gives some
additional information about the error. This is left empty by default.}

\item{n}{Optional. A positive integer which indicates how many invalid
items at most to display. The default value is \code{5}.}

\item{...}{Optional. Additional arguments passed to \code{\link[rlang:abort]{rlang::abort()}},
which is called internally.}
}
\value{
An invisible \code{NULL} if the argument is valid, or an error message
is generated.
}
\description{
Check if each item of an argument has valid type,
and if not, generate an error message.
}
\details{
The R expressions in argument \code{specific} can contain
\itemize{
\item argument \code{name},
\item argument \code{i} which indicates the index of an invalid item, and
\item argument \code{type} which indicates the item's type.
}
}
\examples{
\dontrun{
# argument to check
arg <- as.list(1:10)

check_types(arg, "character")

# specify argument `specific`
check_types(
  arg,
  "character",
  specific = "`{name}[[{i}]]` is an {type}, oh my god!"
)

# change argument `n`
check_types(arg, "character", n = 3)
}
}
\seealso{
"Examples" section in \code{\link[=check_type]{check_type()}} for how to customize error
message and how to add additional arguments.

\code{\link[glue:glue]{glue::glue()}} for inserting R expressions into characters.

\code{\link[=Statement]{Statement()}} for more details about arguments \code{general},
\code{specifics} and \code{supplement}.

\code{\link[rlang:abort]{rlang::abort()}} for adding additional arguments.

\code{vignette("erify")} for a gentle introduction to this package.

\href{https://style.tidyverse.org/error-messages.html}{The tidyverse style guide}
for more details about the used error message style.
}
