% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{join}
\alias{join}
\title{Connect Words with Conjunction}
\usage{
join(words, conjunction = "or")
}
\arguments{
\item{words}{A vector of list whose items can be converted to characters.}

\item{conjunction}{A single character which represents a conjunction word.
The default value is \code{"or"}.}
}
\value{
If has length 1 or less, \code{words} is returned. Or items of \code{words}
are concatenated and returned.
}
\description{
Connect given words with a conjunction, e.g. "and" and
"or".
}
\examples{
words <- c("apple", "orange", "Pink Floyd")
join(words, "and")
}
