% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statement.R
\name{shorten}
\alias{shorten}
\title{Shorten Message}
\usage{
shorten(specifics, n = NULL)
}
\arguments{
\item{specifics}{A character vector which gives a list of message
details.}

\item{n}{Optional. A positive integer which indicates how many message
details at most to display. The default value is \code{5}.}
}
\value{
A character vector.
}
\description{
Shorten a list of message details.
}
\examples{
ss <- letters[1:6]

shorten(ss)

shorten(ss, 3)

shorten(ss, 7)
}
\seealso{
Usually used with \code{\link[=Statement]{Statement()}} and \code{\link[=trigger]{trigger()}}.
}
