% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/where.R
\name{where}
\alias{where}
\alias{is_rmd}
\alias{is_rstudio}
\alias{is_jupyter}
\title{Detect Where Code Is Running}
\usage{
where()

is_rmd()

is_rstudio()

is_jupyter()
}
\value{
For \code{where()}:
\itemize{
\item If executed in R Markdown file, it returns the output format.
If output format is not specified, it returns \code{"rmd"}.
\item If executed in RStudio, it returns \code{"rstudio"}.
\item If executed in R Jupyter Notebook, it returns \code{"jupyter"}.
\item If executed in other contexts, it returns \code{"other"}.
}

\code{is_rmd()}, \code{is_rstudio()} and \code{is_jupyter()} return \code{TRUE} if executed
in their corresponding contexts, or \code{FALSE} if not.
}
\description{
Check if code is running in RStudio, R Markdown file,
R Jupyter Notebook or other contexts. And if is in R Markdown file,
check the output format.
}
\seealso{
\code{\link[rstudioapi:isAvailable]{rstudioapi::isAvailable()}}, \code{\link[knitr:output_type]{knitr::pandoc_to()}}.
}
