% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statement.R
\name{throw}
\alias{throw}
\title{Shortcut to Trigger Message}
\usage{
throw(general, specifics = NULL, env = NULL, n = NULL, as = NULL, ...)
}
\arguments{
\item{general}{A single character which gives a general statement
of a message.}

\item{specifics}{Optional. A character vector which gives a list of details
of a message. If is \code{character(0)}, \code{throw()} returns silently.}

\item{env}{Optional. An environment or named list which is used to evaluate
the R code in the above arguments.}

\item{n}{Optional. A positive integer which indicates how many message
details at most to display. The default value is \code{5}.}

\item{as}{Optional. \code{"error"}, \code{"warning"} or \code{"message"} which indicates
how to trigger the message. The default value is \code{"error"}.}

\item{...}{Optional. Additional arguments which can be passed to
\code{\link[rlang:abort]{rlang::abort()}} or related functions.}
}
\value{
If \code{specifics} is \code{character(0)}, returns an invisible \code{NULL}.
Or generate a normal, warning, or error message.
}
\description{
A shortcut to trigger a message.
}
\examples{
\dontrun{
general <- "You are wrong."
throw(general)

# returns silently if `specifics` is `character(0)`
throw(general, character(0))

# display less details
throw(general, letters, n = 3)
}
}
\seealso{
This function is a shortcut of \code{\link[=shorten]{shorten()}}, \code{\link[=Statement]{Statement()}} and
\code{\link[=trigger]{trigger()}}.
}
